Moved Speaker Code to seperate File than auto.
