+++ src/Components/auto-page/Auto.jsx
@@ -1,139 +1,10 @@
-import * as React from 'react';
-import { Unstable_NumberInput as BaseNumberInput } from '@mui/base/Unstable_NumberInput';
-import { styled } from '@mui/system';
+import React from "react";
+import './Auto.scss'
 
-import './Auto.scss';
-
-import Header from './AutoSpeaker';
-import RemoveIcon from '@mui/icons-material/Remove';
-import AddIcon from '@mui/icons-material/Add';
-
-//-------------------------------------------------------------//
-//      Ronan's Code for Auto Speaker,  God Save us All        //
-//-------------------------------------------------------------//
-
-const NumberInput = React.forwardRef(function CustomNumberInput(props, ref) {
-  return (
-    <BaseNumberInput
-      slots={{
-        root: StyledInputRoot,
-        input: StyledInput,
-
-        incrementButton: StyledButton,
-        decrementButton: StyledButton,
-      }}
-      slotProps={{
-        incrementButton: {
-          children: <AddIcon fontSize="small" />,
-          className: 'increment',
-        },
-
-        decrementButton: {
-          children: <RemoveIcon fontSize="small" />,
-        },
-      }}
-      {...props}
-      ref={ref}
-    />
-  );
+export default function Auto() {
+    return (
+        <>
+        <h1>test</h1>
+        </>
+    )
 }
-
-//---------------------//
-// C O L O R S ! ! ! ! //
-// Y I P P P P E E E E //
-//---------------------//
-
-);
-
-export default function QuantityInput() {
-  return <NumberInput aria-label="Quantity Input" min={0} max={199} defaultValue="0"/>;
-}
-
-const blue = '#02243e';
-
-const grey = '#babfb7';
-
-const StyledInputRoot = styled('div')(
-  ({ theme }) => `
-  font-family: 'IBM Plex Sans', sans-serif;
-  font-weight: 400;
-  color: ${theme.palette.mode === 'dark' ? grey : grey};
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: center;
-  align-items: center;
-`,
-);
-
-const StyledInput = styled('input')(
-  ({ theme }) => `
-  font-size: 0.875rem;
-  font-family: inherit;
-  font-weight: 400;
-  line-height: 1.375;
-  color: ${theme.palette.mode === 'dark' ? grey : grey};
-  background: ${theme.palette.mode === 'dark' ? grey : '#fff'};
-  border: 1px solid ${theme.palette.mode === 'dark' ? grey : grey};
-  box-shadow: 0px 2px 4px ${
-    theme.palette.mode === 'dark' ? 'rgba(0,0,0, 0.5)' : 'rgba(0,0,0, 0.05)'
-  };
-  border-radius: 8px;
-  margin: 0 8px;
-  padding: 10px 12px;
-  outline: 0;
-  min-width: 0;
-  width: 4rem;
-  text-align: center;
-
-  &:hover {
-    border-color: ${blue};
-  }
-
-  &:focus {
-    border-color: ${blue};
-    box-shadow: 0 0 0 3px ${theme.palette.mode === 'dark' ? blue : blue};
-  }
-
-  &:focus-visible {
-    outline: 0;
-  }
-`,
-);
-
-const StyledButton = styled('button')(
-  ({ theme }) => `
-  font-family: 'IBM Plex Sans', sans-serif;
-  font-size: 0.875rem;
-  box-sizing: border-box;
-  line-height: 1.5;
-  border: 1px solid;
-  border-radius: 999px;
-  border-color: ${theme.palette.mode === 'dark' ? grey : grey};
-  background: ${theme.palette.mode === 'dark' ? grey : grey};
-  color: ${theme.palette.mode === 'dark' ? grey : grey};
-  width: 32px;
-  height: 32px;
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: center;
-  align-items: center;
-  transition-property: all;
-  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
-  transition-duration: 120ms;
-
-  &:hover {
-    cursor: pointer;
-    background: ${theme.palette.mode === 'dark' ? blue : blue};
-    border-color: ${theme.palette.mode === 'dark' ? blue : blue};
-    color: ${grey};
-  }
-
-  &:focus-visible {
-    outline: 0;
-  }
-
-  &.increment {
-    order: 1;
-  }
-`,
-);
